+++ bin/wasm-node/javascript/bindings-wasi.js
